Reorganization of the test folder
Basic programming concepts were used and responsibilities were divided into more manageable groups in terms of size.
