+++ src/test/java/co/edu/unbosque/connection/model/persistence/GeneralServiceDAOForTest.java
@@ -1,102 +1,105 @@
-package co.edu.unbosque.model.persistence;
+package co.edu.unbosque.connection.model.persistence;
 
 import java.sql.Date;
 import java.sql.SQLException;
 import java.time.LocalDate;
 import java.util.ArrayList;
-import co.edu.unbosque.controller.DBConnectionForTest;
-import co.edu.unbosque.model.ParticipantDTO;
 
-public class ParticipantDAOForTest implements CRUDoperation<ParticipantDTO> {
+import co.edu.unbosque.connection.controller.DBConnectionForTest;
+import co.edu.unbosque.model.GeneralServiceDTO;
+import co.edu.unbosque.model.persistence.CRUDoperation;
 
-	private ArrayList<ParticipantDTO> listParticipants;
+public class GeneralServiceDAOForTest implements CRUDoperation<GeneralServiceDTO> {
+
+	private ArrayList<GeneralServiceDTO> listGeneralService;
 	private DBConnectionForTest dbcon;
 
 	/**
-	 * Create an object of class ParticipantDAO
+	 * Create an object of class GeneralServiceDAO.
 	 */
-	public ParticipantDAOForTest() {
+	public GeneralServiceDAOForTest() {
 		dbcon = new DBConnectionForTest();
-		listParticipants = loadData();
-		if (listParticipants == null)
-			listParticipants = new ArrayList<>();
+		listGeneralService = loadData();
+		if (listGeneralService == null)
+			listGeneralService = new ArrayList<>();
 	}
 
 	/**
-	 * Gets the ArrayList containing ParticipantDTO objects.
+	 * Gets the ArrayList containing GeneralServiceDTO objects.
 	 * 
-	 * @return The ArrayList of ParticipantDTO objects.
+	 * @return The ArrayList of PersonDTO objects.
 	 */
-	public ArrayList<ParticipantDTO> getListParticipants() {
-		return listParticipants;
+	public ArrayList<GeneralServiceDTO> getListGeneralService() {
+		return listGeneralService;
 	}
 
 	/**
-	 * Update the ArrayList containing ParticipantDTO objects.
+	 * Update the ArrayList containing GeneralServiceDTO objects.
 	 * 
-	 * @param listParticipants :The ArrayList of ParticipantDTO objects.
+	 * @param listGeneralService :The ArrayList of GeneralServiceDTO objects.
 	 */
-	public void setListParticipants(ArrayList<ParticipantDTO> listParticipants) {
-		this.listParticipants = listParticipants;
+	public void setListGeneralService(ArrayList<GeneralServiceDTO> listGeneralService) {
+		this.listGeneralService = listGeneralService;
 	}
 
 	@Override
-	public void create(ParticipantDTO info) {
-		dbcon.initConnection();
+	public void create(GeneralServiceDTO info) {
 		try {
+			dbcon.initConnection();
 			dbcon.setPreparedstatement(dbcon.getConnect().prepareStatement(
-					"INSERT INTO participants (username, userpassword, personname, document, dateofbirth, cityofbirth, participatedsessions, nickname) VALUES(?,?,?,?,?,?,?,?);"));
+					"INSERT INTO generalservice (username, userpassword, personname, document, dateofbirth, cityofbirth, salary, sessionscleaned) VALUES(?,?,?,?,?,?,?,?);"));
 			dbcon.getPreparedstatement().setString(1, info.getUserName());
 			dbcon.getPreparedstatement().setString(2, info.getPassword());
 			dbcon.getPreparedstatement().setString(3, info.getName());
 			dbcon.getPreparedstatement().setLong(4, info.getDocument());
 			dbcon.getPreparedstatement().setDate(5, Date.valueOf(info.getDateOfBirth()));
 			dbcon.getPreparedstatement().setString(6, info.getCityOfBirth());
-			dbcon.getPreparedstatement().setInt(7, info.getParticipatedSessions());
-			dbcon.getPreparedstatement().setString(8, info.getNickName());
+			dbcon.getPreparedstatement().setDouble(7, info.getSalary());
+			dbcon.getPreparedstatement().setInt(8, info.getSessionsCleaned());
 			dbcon.getPreparedstatement().executeUpdate();
 			dbcon.close();
-			listParticipants.add(info);
+			listGeneralService.add(info);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
+
 	}
 
 	@Override
-	public void update(int index, ParticipantDTO info) {
-		long adocument = listParticipants.get(index).getDocument();
+	public void update(int index, GeneralServiceDTO info) {
+		long adocument = listGeneralService.get(index).getDocument();
 		dbcon.initConnection();
 		try {
 			dbcon.setPreparedstatement(dbcon.getConnect().prepareStatement(
-					"UPDATE participants SET username=?, userpassword=?, personname=?, document=?, dateofbirth=?, cityofbirth=?, participatedsessions=?, nickname=? WHERE document=?;"));
+					"UPDATE generalservice SET username=?, userpassword=?, personname=?, document=?, dateofbirth=?, cityofbirth=?, salary=?, sessionscleaned=? WHERE document=?;"));
 			dbcon.getPreparedstatement().setString(1, info.getUserName());
 			dbcon.getPreparedstatement().setString(2, info.getPassword());
 			dbcon.getPreparedstatement().setString(3, info.getName());
 			dbcon.getPreparedstatement().setLong(4, info.getDocument());
 			dbcon.getPreparedstatement().setDate(5, Date.valueOf(info.getDateOfBirth()));
 			dbcon.getPreparedstatement().setString(6, info.getCityOfBirth());
-			dbcon.getPreparedstatement().setInt(7, info.getParticipatedSessions());
-			dbcon.getPreparedstatement().setString(8, info.getNickName());
+			dbcon.getPreparedstatement().setDouble(7, info.getSalary());
+			dbcon.getPreparedstatement().setInt(8, info.getSessionsCleaned());
 			dbcon.getPreparedstatement().setLong(9, adocument);
 			dbcon.getPreparedstatement().executeUpdate();
 			dbcon.close();
-			listParticipants.set(index, info);
+			listGeneralService.set(index, info);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 	}
 
 	@Override
-	public ParticipantDTO delete(int index) {
-		long document = listParticipants.get(index).getDocument();
+	public GeneralServiceDTO delete(int index) {
+		long document = listGeneralService.get(index).getDocument();
 		dbcon.initConnection();
 		try {
 			dbcon.setPreparedstatement(
-					dbcon.getConnect().prepareStatement("DELETE FROM participants WHERE document=?;"));
+					dbcon.getConnect().prepareStatement("DELETE FROM generalservice WHERE document=?;"));
 			dbcon.getPreparedstatement().setLong(1, document);
 			dbcon.getPreparedstatement().executeUpdate();
 			dbcon.close();
-			return listParticipants.remove(index);
+			return listGeneralService.remove(index);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
@@ -106,15 +109,15 @@ public class ParticipantDAOForTest implements CRUDoperation<ParticipantDTO> {
 	/**
 	 * Loads the information from the database.
 	 * 
-	 * @return The ArrayList of ParticipantDTO objects with information from the
+	 * @return The ArrayList of GeneralServiceDTO objects with information from the
 	 *         database.
 	 */
-	public ArrayList<ParticipantDTO> loadData() {
-		ArrayList<ParticipantDTO> data = new ArrayList<>();
+	public ArrayList<GeneralServiceDTO> loadData() {
+		ArrayList<GeneralServiceDTO> data = new ArrayList<>();
 		dbcon.initConnection();
 		try {
 			dbcon.setStatement(dbcon.getConnect().createStatement());
-			dbcon.setResultset(dbcon.getStatement().executeQuery("SELECT *FROM participants;"));
+			dbcon.setResultset(dbcon.getStatement().executeQuery("SELECT *FROM generalservice;"));
 			while (dbcon.getResultset().next()) {
 				String userName = dbcon.getResultset().getString("username");
 				String password = dbcon.getResultset().getString("userpassword");
@@ -122,10 +125,10 @@ public class ParticipantDAOForTest implements CRUDoperation<ParticipantDTO> {
 				long document = dbcon.getResultset().getLong("document");
 				LocalDate dateOfBirth = dbcon.getResultset().getDate("dateofbirth").toLocalDate();
 				String cityOfBirth = dbcon.getResultset().getString("cityofbirth");
-				int participatedSessions = dbcon.getResultset().getInt("participatedsessions");
-				String nickName = dbcon.getResultset().getString("nickname");
-				data.add(new ParticipantDTO(userName, password, name, document, dateOfBirth, cityOfBirth,
-						participatedSessions, nickName));
+				double salary = dbcon.getResultset().getDouble("salary");
+				int sessionsCleaned = dbcon.getResultset().getInt("sessionscleaned");
+				data.add(new GeneralServiceDTO(userName, password, name, document, dateOfBirth, cityOfBirth, salary,
+						sessionsCleaned));
 			}
 			dbcon.close();
 			return data;
@@ -143,8 +146,8 @@ public class ParticipantDAOForTest implements CRUDoperation<ParticipantDTO> {
 	 * @return True if the username and password exist, False if not.
 	 */
 	public boolean containsUser(String username, String password) {
-		for (ParticipantDTO partcipant : listParticipants) {
-			if (partcipant.getUserName().equals(username) && partcipant.getPassword().equals(password)) {
+		for (GeneralServiceDTO service : listGeneralService) {
+			if (service.getUserName().equals(username) && service.getPassword().equals(password)) {
 				return true;
 			}
 		}
@@ -158,8 +161,8 @@ public class ParticipantDAOForTest implements CRUDoperation<ParticipantDTO> {
 	 * @return True if the username exist, False if not.
 	 */
 	public boolean containsUserName(String username) {
-		for (ParticipantDTO partcipant : listParticipants) {
-			if (partcipant.getUserName().equals(username)) {
+		for (GeneralServiceDTO service : listGeneralService) {
+			if (service.getUserName().equals(username)) {
 				return true;
 			}
 		}
@@ -173,11 +176,12 @@ public class ParticipantDAOForTest implements CRUDoperation<ParticipantDTO> {
 	 * @return True if the password exist, False if not.
 	 */
 	public boolean containsPassword(String password) {
-		for (ParticipantDTO partcipant : listParticipants) {
-			if (partcipant.getPassword().equals(password)) {
+		for (GeneralServiceDTO service : listGeneralService) {
+			if (service.getPassword().equals(password)) {
 				return true;
 			}
 		}
 		return false;
 	}
+
 }
