+++ src/test/java/co/edu/unbosque/connection/model/persistence/ParticipantDAOForTest.java
@@ -1,103 +1,104 @@
-package co.edu.unbosque.model.persistence;
+package co.edu.unbosque.connection.model.persistence;
 
 import java.sql.Date;
 import java.sql.SQLException;
 import java.time.LocalDate;
 import java.util.ArrayList;
-import co.edu.unbosque.controller.DBConnectionForTest;
-import co.edu.unbosque.model.GeneralServiceDTO;
 
-public class GeneralServiceDAOForTest implements CRUDoperation<GeneralServiceDTO> {
+import co.edu.unbosque.connection.controller.DBConnectionForTest;
+import co.edu.unbosque.model.ParticipantDTO;
+import co.edu.unbosque.model.persistence.CRUDoperation;
 
-	private ArrayList<GeneralServiceDTO> listGeneralService;
+public class ParticipantDAOForTest implements CRUDoperation<ParticipantDTO> {
+
+	private ArrayList<ParticipantDTO> listParticipants;
 	private DBConnectionForTest dbcon;
 
 	/**
-	 * Create an object of class GeneralServiceDAO.
+	 * Create an object of class ParticipantDAO
 	 */
-	public GeneralServiceDAOForTest() {
+	public ParticipantDAOForTest() {
 		dbcon = new DBConnectionForTest();
-		listGeneralService = loadData();
-		if (listGeneralService == null)
-			listGeneralService = new ArrayList<>();
+		listParticipants = loadData();
+		if (listParticipants == null)
+			listParticipants = new ArrayList<>();
 	}
 
 	/**
-	 * Gets the ArrayList containing GeneralServiceDTO objects.
+	 * Gets the ArrayList containing ParticipantDTO objects.
 	 * 
-	 * @return The ArrayList of PersonDTO objects.
+	 * @return The ArrayList of ParticipantDTO objects.
 	 */
-	public ArrayList<GeneralServiceDTO> getListGeneralService() {
-		return listGeneralService;
+	public ArrayList<ParticipantDTO> getListParticipants() {
+		return listParticipants;
 	}
 
 	/**
-	 * Update the ArrayList containing GeneralServiceDTO objects.
+	 * Update the ArrayList containing ParticipantDTO objects.
 	 * 
-	 * @param listGeneralService :The ArrayList of GeneralServiceDTO objects.
+	 * @param listParticipants :The ArrayList of ParticipantDTO objects.
 	 */
-	public void setListGeneralService(ArrayList<GeneralServiceDTO> listGeneralService) {
-		this.listGeneralService = listGeneralService;
+	public void setListParticipants(ArrayList<ParticipantDTO> listParticipants) {
+		this.listParticipants = listParticipants;
 	}
 
 	@Override
-	public void create(GeneralServiceDTO info) {
-		try {
-			dbcon.initConnection();
-			dbcon.setPreparedstatement(dbcon.getConnect().prepareStatement(
-					"INSERT INTO generalservice (username, userpassword, personname, document, dateofbirth, cityofbirth, salary, sessionscleaned) VALUES(?,?,?,?,?,?,?,?);"));
-			dbcon.getPreparedstatement().setString(1, info.getUserName());
-			dbcon.getPreparedstatement().setString(2, info.getPassword());
-			dbcon.getPreparedstatement().setString(3, info.getName());
-			dbcon.getPreparedstatement().setLong(4, info.getDocument());
-			dbcon.getPreparedstatement().setDate(5, Date.valueOf(info.getDateOfBirth()));
-			dbcon.getPreparedstatement().setString(6, info.getCityOfBirth());
-			dbcon.getPreparedstatement().setDouble(7, info.getSalary());
-			dbcon.getPreparedstatement().setInt(8, info.getSessionsCleaned());
-			dbcon.getPreparedstatement().executeUpdate();
-			dbcon.close();
-			listGeneralService.add(info);
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
-
-	}
-
-	@Override
-	public void update(int index, GeneralServiceDTO info) {
-		long adocument = listGeneralService.get(index).getDocument();
+	public void create(ParticipantDTO info) {
 		dbcon.initConnection();
 		try {
 			dbcon.setPreparedstatement(dbcon.getConnect().prepareStatement(
-					"UPDATE generalservice SET username=?, userpassword=?, personname=?, document=?, dateofbirth=?, cityofbirth=?, salary=?, sessionscleaned=? WHERE document=?;"));
+					"INSERT INTO participants (username, userpassword, personname, document, dateofbirth, cityofbirth, participatedsessions, nickname) VALUES(?,?,?,?,?,?,?,?);"));
 			dbcon.getPreparedstatement().setString(1, info.getUserName());
 			dbcon.getPreparedstatement().setString(2, info.getPassword());
 			dbcon.getPreparedstatement().setString(3, info.getName());
 			dbcon.getPreparedstatement().setLong(4, info.getDocument());
 			dbcon.getPreparedstatement().setDate(5, Date.valueOf(info.getDateOfBirth()));
 			dbcon.getPreparedstatement().setString(6, info.getCityOfBirth());
-			dbcon.getPreparedstatement().setDouble(7, info.getSalary());
-			dbcon.getPreparedstatement().setInt(8, info.getSessionsCleaned());
+			dbcon.getPreparedstatement().setInt(7, info.getParticipatedSessions());
+			dbcon.getPreparedstatement().setString(8, info.getNickName());
+			dbcon.getPreparedstatement().executeUpdate();
+			dbcon.close();
+			listParticipants.add(info);
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+	}
+
+	@Override
+	public void update(int index, ParticipantDTO info) {
+		long adocument = listParticipants.get(index).getDocument();
+		dbcon.initConnection();
+		try {
+			dbcon.setPreparedstatement(dbcon.getConnect().prepareStatement(
+					"UPDATE participants SET username=?, userpassword=?, personname=?, document=?, dateofbirth=?, cityofbirth=?, participatedsessions=?, nickname=? WHERE document=?;"));
+			dbcon.getPreparedstatement().setString(1, info.getUserName());
+			dbcon.getPreparedstatement().setString(2, info.getPassword());
+			dbcon.getPreparedstatement().setString(3, info.getName());
+			dbcon.getPreparedstatement().setLong(4, info.getDocument());
+			dbcon.getPreparedstatement().setDate(5, Date.valueOf(info.getDateOfBirth()));
+			dbcon.getPreparedstatement().setString(6, info.getCityOfBirth());
+			dbcon.getPreparedstatement().setInt(7, info.getParticipatedSessions());
+			dbcon.getPreparedstatement().setString(8, info.getNickName());
 			dbcon.getPreparedstatement().setLong(9, adocument);
 			dbcon.getPreparedstatement().executeUpdate();
 			dbcon.close();
-			listGeneralService.set(index, info);
+			listParticipants.set(index, info);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 	}
 
 	@Override
-	public GeneralServiceDTO delete(int index) {
-		long document = listGeneralService.get(index).getDocument();
+	public ParticipantDTO delete(int index) {
+		long document = listParticipants.get(index).getDocument();
 		dbcon.initConnection();
 		try {
 			dbcon.setPreparedstatement(
-					dbcon.getConnect().prepareStatement("DELETE FROM generalservice WHERE document=?;"));
+					dbcon.getConnect().prepareStatement("DELETE FROM participants WHERE document=?;"));
 			dbcon.getPreparedstatement().setLong(1, document);
 			dbcon.getPreparedstatement().executeUpdate();
 			dbcon.close();
-			return listGeneralService.remove(index);
+			return listParticipants.remove(index);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
@@ -107,15 +108,15 @@ public class GeneralServiceDAOForTest implements CRUDoperation<GeneralServiceDTO
 	/**
 	 * Loads the information from the database.
 	 * 
-	 * @return The ArrayList of GeneralServiceDTO objects with information from the
+	 * @return The ArrayList of ParticipantDTO objects with information from the
 	 *         database.
 	 */
-	public ArrayList<GeneralServiceDTO> loadData() {
-		ArrayList<GeneralServiceDTO> data = new ArrayList<>();
+	public ArrayList<ParticipantDTO> loadData() {
+		ArrayList<ParticipantDTO> data = new ArrayList<>();
 		dbcon.initConnection();
 		try {
 			dbcon.setStatement(dbcon.getConnect().createStatement());
-			dbcon.setResultset(dbcon.getStatement().executeQuery("SELECT *FROM generalservice;"));
+			dbcon.setResultset(dbcon.getStatement().executeQuery("SELECT *FROM participants;"));
 			while (dbcon.getResultset().next()) {
 				String userName = dbcon.getResultset().getString("username");
 				String password = dbcon.getResultset().getString("userpassword");
@@ -123,10 +124,10 @@ public class GeneralServiceDAOForTest implements CRUDoperation<GeneralServiceDTO
 				long document = dbcon.getResultset().getLong("document");
 				LocalDate dateOfBirth = dbcon.getResultset().getDate("dateofbirth").toLocalDate();
 				String cityOfBirth = dbcon.getResultset().getString("cityofbirth");
-				double salary = dbcon.getResultset().getDouble("salary");
-				int sessionsCleaned = dbcon.getResultset().getInt("sessionscleaned");
-				data.add(new GeneralServiceDTO(userName, password, name, document, dateOfBirth, cityOfBirth, salary,
-						sessionsCleaned));
+				int participatedSessions = dbcon.getResultset().getInt("participatedsessions");
+				String nickName = dbcon.getResultset().getString("nickname");
+				data.add(new ParticipantDTO(userName, password, name, document, dateOfBirth, cityOfBirth,
+						participatedSessions, nickName));
 			}
 			dbcon.close();
 			return data;
@@ -144,8 +145,8 @@ public class GeneralServiceDAOForTest implements CRUDoperation<GeneralServiceDTO
 	 * @return True if the username and password exist, False if not.
 	 */
 	public boolean containsUser(String username, String password) {
-		for (GeneralServiceDTO service : listGeneralService) {
-			if (service.getUserName().equals(username) && service.getPassword().equals(password)) {
+		for (ParticipantDTO partcipant : listParticipants) {
+			if (partcipant.getUserName().equals(username) && partcipant.getPassword().equals(password)) {
 				return true;
 			}
 		}
@@ -159,8 +160,8 @@ public class GeneralServiceDAOForTest implements CRUDoperation<GeneralServiceDTO
 	 * @return True if the username exist, False if not.
 	 */
 	public boolean containsUserName(String username) {
-		for (GeneralServiceDTO service : listGeneralService) {
-			if (service.getUserName().equals(username)) {
+		for (ParticipantDTO partcipant : listParticipants) {
+			if (partcipant.getUserName().equals(username)) {
 				return true;
 			}
 		}
@@ -174,12 +175,11 @@ public class GeneralServiceDAOForTest implements CRUDoperation<GeneralServiceDTO
 	 * @return True if the password exist, False if not.
 	 */
 	public boolean containsPassword(String password) {
-		for (GeneralServiceDTO service : listGeneralService) {
-			if (service.getPassword().equals(password)) {
+		for (ParticipantDTO partcipant : listParticipants) {
+			if (partcipant.getPassword().equals(password)) {
 				return true;
 			}
 		}
 		return false;
 	}
-
 }
