+++ src/test/java/co/edu/unbosque/connection/model/persistence/PsychologistDAOForTest.java
@@ -1,4 +1,4 @@
-package co.edu.unbosque.model.persistence;
+package co.edu.unbosque.connection.model.persistence;
 
 import java.sql.Date;
 import java.sql.SQLException;
@@ -6,9 +6,10 @@ import java.time.LocalDate;
 import java.time.Year;
 import java.util.ArrayList;
 
+import co.edu.unbosque.connection.controller.DBConnectionForTest;
 import co.edu.unbosque.controller.DBConnection;
-import co.edu.unbosque.controller.DBConnectionForTest;
 import co.edu.unbosque.model.PsychologistDTO;
+import co.edu.unbosque.model.persistence.CRUDoperation;
 
 public class PsychologistDAOForTest implements CRUDoperation<PsychologistDTO> {
